backend/auth_service/src/requirements.txt
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
redis==4.4.2
alembic==1.13.1
pydantic==2.5.2
pydantic-settings==2.1.0
uvicorn==0.28.0
uvloop==0.19.0 ; sys_platform != "win32" and implementation_name == "cpython"
orjson==3.9.10
fastapi==0.109.0
typer==0.9.0
werkzeug==3.0.1
alembic==1.13.1
python-jose[cryptography]==3.3.0
grpcio-tools==1.62.1
aiohttp==3.9.0b0
opentelemetry-api==1.17.0
opentelemetry-sdk==1.17.0
opentelemetry-instrumentation-fastapi==0.38b0
opentelemetry-exporter-jaeger==1.17.0
fastapi_limiter==0.1.6
asgi-correlation-id==4.3.1